make m1n1 package compatible with new way to package things

+++ gpu.sh
@@ -54,7 +54,7 @@ build_uboot()
 package_boot_bin()
 {
 (
-        export M1N1_VERSION=1.2.3-1
+        export M1N1_VERSION=1.2.3-2
         rm -rf m1n1_${M1N1_VERSION}_arm64
         mkdir -p m1n1_${M1N1_VERSION}_arm64/DEBIAN m1n1_${M1N1_VERSION}_arm64/usr/lib/m1n1/
         cp u-boot.bin m1n1_${M1N1_VERSION}_arm64/usr/lib/m1n1/boot.bin
@@ -73,7 +73,10 @@ EOF
 #!/bin/bash
 
 export PATH=/bin
-cp /boot/efi/m1n1/boot.bin /boot/efi/m1n1/`date +%Y%m%d%H%M`.bin
+if [ -f /boot/efi/m1n1/boot.bin ]; then
+        cp /boot/efi/m1n1/boot.bin /boot/efi/m1n1/`date +%Y%m%d%H%M`.bin
+fi
+mkdir -p /boot/efi/m1n1/
 cp /usr/lib/m1n1/boot.bin /boot/efi/m1n1/
 EOF
